refactor(tech): move logger constants to shared context
